fix annotation problem

+++ 0x01-python_async_function/1-concurrent_coroutines.py
@@ -7,7 +7,7 @@ import asyncio
 wait_random = __import__('0-basic_async_syntax').wait_random
 
 
-async def wait_n(n: float, max_delay: float) -> float:
+async def wait_n(n: int, max_delay: int) -> float:
     """the function that generate a random number and save them in the list"""
 
     tasks = [asyncio.create_task(wait_random(max_delay)) for _ in range(n)]
